requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.42.0
django-allauth-bootstrap==0.2
django-bootstrap-form==3.4
django-urls==1.1.3
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
